Refactor discovery and clean up codebase
What

This is a massive change that completely overhauls discovery, selectors
and filtering.

It also updates the behaviour of discovery to ignore subresources and
filter-out API groups which are not supposed to be edited by API
clients.

It also reshuffles the codebase to put some structure to it.

Why

To address multiple bugs related to discovery.

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>