Remove server-side fields when pulling resources
What

This commit changes the logic of pulling resources to strip them from
server-side fields like resourceVersion, generation and so on, as well
as some well-known server-side annotations & labels.

It also removes the "overwrite" flag used for pushing since that is the
new default behaviour.

Why

We are removing the ability to pull resources in their "backup" format
for now since it prevents the use-cases of e.g. migrating resources from
one instance to another and is overall a buggy experience.

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>