Fix CI security scan results
I ran `zimzor` against the repo and wanted to address the high
vulnernabilities that could easily be exploited. This addresses some of
the more likely scenarios and updated permissions to be scoped to
specific steps. With CI disabled, these are difficult to guarantee
success, but gets us closer to a better state and should be relatively
easy to fix in the future.