Replace pull / push / validate directories with paths
What

This commit changes how `resources pull / push / validate` interact with
the FS.

Rather than working with directories only, they now all work with
arbitrary paths (which could be either directories or files).

That said, `resources push` will expect the `--path` argument to be a
directory (and will error out otherwise) because we do not support
writing (or reading) multiple resources into one file at the moment.

However, for the sake of consistency it uses the same option name.

Why

To make it possible to push & validate individual files instead of just
directories.

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>