Support paginated responses in list requests
What

This commit updates the List method for the clients to automatically
fetch all objects, even if the response is paginated.

This is done by using the built-in `client-go` pager.

This commit also updates the `GetMultiple` method on the clients to use
concurrent `Get` requests under the hood, rather than relying on the
`List`.

Lastly, it fixes a bug where fetching all objects of a given API group
was sending duplicate list requests for each API versions.

Why

Long responses for List requests can be paginated by the server, meaning
that the CLI won't actually fetch everything, instead only fetching the
first batch of objects.

The changes to GetMultiple were required for the same reason - in the
presence of pagination, using List is very wasteful.

The fix for fetching duplicate objects via multiple API versions is
aimed to fixing performance issues.

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>