Automatically discover `stack-id` (#164)
We don't need users to specify `stack-id` - we can figure it out on the
fly. IF its specified, we will check whether it matches what we
discover.

It's kept around in config - if the autodiscovery starts failing for
some reason, then setting `stack-id` can be used as a temporary
workaround